Added automatic detection of Motion Direction
Motion Direction, which is needed for acquisition.xml and correct IK down the line, is now (with some excellent advice from Claire Crossley) detected automatically through two new tags (CoordinateSystemOrientation & PelvisMarkers) in the autoC3Dsetup.xml. The order of PelvisMarkers is crucial for correct function, which has been updated and highlited in the User Guide.docx.